ajout de params sur les outils de dev

+++ tools/devreset.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 
-rm -rf ./app/logs/;
-rm -rf ./app/cache/;
+if [ $1 -eq 'cache' ]; then
+    rm -rf ./app/logs/;
+    rm -rf ./app/cache/;
 
-php ./app/console cache:clear --env=dev;
-php ./app/console assetic:dump --env=dev;
-php ./app/console assets:install --env=dev;
-chmod -R 777 ./app/logs/;
-chmod -R 777 ./app/cache/;
+    php ./app/console cache:clear --env=dev;
+    php ./app/console assetic:dump --env=dev;
+    php ./app/console assets:install --env=dev;
+    chmod -R 777 ./app/logs/;
+    chmod -R 777 ./app/cache/;
+fi
 
-php app/console doctrine:database:drop --env=dev --force
-php app/console doctrine:database:create --env=dev
-php app/console doctrine:schema:create --env=dev
+if [$1 -eq 'db' ]; then
+    php app/console doctrine:database:drop --env=dev --force
+    php app/console doctrine:database:create --env=dev
+    php app/console doctrine:schema:create --env=dev
 
-php app/console doctrine:fixtures:load --no-interaction
+    php app/console doctrine:fixtures:load --no-interaction
+fi
